requirements.txt
imagenet-c==0.0.3
matplotlib==3.4.1
numpy==1.21.0
opencv-python
opencv-python-headless
pandas==1.2.5
pickleshare==0.7.5
Pillow==8.3.0
pyDeprecate==0.3.0
pyparsing==2.4.7
pytorch-lightning==1.3.0
rawpy==0.16.0
scikit-image==0.18.1
scipy==1.7.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.4.8
tqdm==4.60.0
future~=0.18.2
jinja2==3.0.1
cma
nevergrad
scikit-learn
scikit-image
lpips